Declassed NoiseLevelEstimation routines and provided appropriate tests
