+++ sunkit_image/utils/noise.py
@@ -1,274 +1,274 @@
-"""
-This module implements a class for noise level estimation.
-"""
+'''
+This module implements a series of functions for noise level estimation
+'''
 
 import numpy as np
 from scipy.stats import gamma
 from skimage.util import view_as_windows
 from scipy.ndimage import correlate
 
-__all__ = ["NoiseLevelEstimation"]
+__all__ = ['noise_estimation','noiselevel','convmtx2','weaktexturemask']
 
+#TODO: Documentation and comments
 
-class NoiseLevelEstimation:
-    def __init__(self, img, patchsize=7, decim=0, conf=1 - 1e-6, itr=3):
-        """
-        Estimates the noise level of an image.
+def noise_estimation(img, patchsize=7, decim=0, conf=1 - 1e-6, itr=3):
 
-        Additive white Gaussian noise (AWGN) is a basic noise model used in Information Theory
-        to mimic the effect of many random processes that occur in nature.
+    """
+    Estimates the noise level of an image.
 
-        Parameters
-        ----------
-        img: `numpy.ndarray`
-            Single Numpy image array.
-        patchsize : `int`, optional
-            Patch size, defaults to 7.
-        decim : `int`, optional
-            Decimation factor, defaults to 0.
-            If you use large number, the calculation will be accelerated.
-        conf : `float`, optional
-            Confidence interval to determine the threshold for the weak texture.
-            In this algorithm, this value is usually set the value very close to one.
-            Defaults to 0.99.
-        itr : `int`, optional
-            Number of iterations,  defaults to 3.
+    Additive white Gaussian noise (AWGN) is a basic noise model used in Information Theory
+    to mimic the effect of many random processes that occur in nature.
 
-        Attributes
-        ----------
-        nlevel: `numpy.ndarray`
-            Estimated noise levels.
-        th: `numpy.ndarray`
-            Threshold to extract weak texture patches at the last iteration.
-        num: `numpy.ndarray`
-            Number of extracted weak texture patches at the last iteration.
-        mask: `numpy.ndarray`
-            Weak-texture mask.
-            0 and 1 represent non-weak-texture and weak-texture regions, respectively.
+    Parameters
+    ----------
+    img: `numpy.ndarray`
+        Single Numpy image array.
+    patchsize : `int`, optional
+        Patch size, defaults to 7.
+    decim : `int`, optional
+        Decimation factor, defaults to 0.
+        If you use large number, the calculation will be accelerated.
+    conf : `float`, optional
+        Confidence interval to determine the threshold for the weak texture.
+        In this algorithm, this value is usually set the value very close to one.
+        Defaults to 0.99.
+    itr : `int`, optional
+        Number of iterations,  defaults to 3.
 
-        Examples
-        --------
-        >>> import numpy as np
-        >>> np.random.seed(0)
-        >>> noisy_image_array = np.random.randn(100, 100)
-        >>> estimate = NoiseLevelEstimation(noisy_image_array, patchsize=11, itr=10)
-        >>> estimate.mask
-        array([[1., 1., 1., ..., 1., 1., 0.],
-            [1., 1., 1., ..., 1., 1., 0.],
-            [1., 1., 1., ..., 1., 1., 0.],
-            ...,
-            [1., 1., 1., ..., 1., 1., 0.],
-            [1., 1., 1., ..., 1., 1., 0.],
-            [0., 0., 0., ..., 0., 0., 0.]])
-        >>> estimate.nlevel
-        array([1.0014616])
-        >>> estimate.th
-        array([173.61530607])
-        >>> estimate.num
-         array([8100.])
+    Returns
+    ----------
+    nlevel: `numpy.ndarray`
+        Estimated noise levels.
+    th: `numpy.ndarray`
+        Threshold to extract weak texture patches at the last iteration.
+    num: `numpy.ndarray`
+        Number of extracted weak texture patches at the last iteration.
+    mask: `numpy.ndarray`
+        Weak-texture mask.
+    0 and 1 represent non-weak-texture and weak-texture regions, respectively.
+    """
 
-        References
-        ----------
-        * Xinhao Liu, Masayuki Tanaka and Masatoshi Okutomi
-          Noise Level Estimation Using Weak Textured Patches of a Single Noisy Image
-          IEEE International Conference on Image Processing (ICIP), 2012.
+    if not(type(img) is np.ndarray):
+        raise TypeError('Input image should be a NumPy ndarray')
 
-        * Xinhao Liu, Masayuki Tanaka and Masatoshi Okutomi
-          Single-Image Noise Level Estimation for Blind Denoising Noisy Image
-          IEEE Transactions on Image Processing, Vol.22, No.12, pp.5226-5237, December, 2013.
-        """
+    if not(type(patchsize) is int):
+        try:
+            patchsize = int(patchsize)
+        except ValueError:
+            raise TypeError('patchsize must be an integer, or int-compatible, variable')
 
-        self.img = img
-        self.patchsize = patchsize
-        self.decim = decim
-        self.conf = conf
-        self.itr = itr
+    if not(type(decim) is int):
+        try:
+            decim = int(decim)
+        except ValueError:
+            raise TypeError('decim must be an integer, or int-compatible, variable')
 
-        self.nlevel, self.th, self.num = self.noiselevel()
-        self.mask = self.weaktexturemask()
+    if not(type(conf) is float):
+        try:
+            conf = float(conf)
+        except ValueError:
+            raise TypeError('conf must be a float, or float-compatible, value between 0 and 1')
 
-    def noiselevel(self):
-        """
-        Calculates the noise level of the input array.
-        """
-        if len(self.img.shape) < 3:
-            self.img = np.expand_dims(self.img, 2)
+    if not(conf >=0 and conf <= 1):
+        raise ValueError('conf must be defined in the interval 0 <= conf <= 1')
+    
+    if not(type(itr) is int):
+        try:
+            itr = int(itr)
+        except ValueError:
+            raise TypeError('itr must be an integer, or int-compatible, variable')
 
-        nlevel = np.ndarray(self.img.shape[2])
-        th = np.ndarray(self.img.shape[2])
-        num = np.ndarray(self.img.shape[2])
+    nlevel, th, num = noiselevel(img, patchsize, decim, conf, itr)
+    mask = weaktexturemask(img, patchsize, th)
 
-        kh = np.expand_dims(np.expand_dims(np.array([-0.5, 0, 0.5]), 0), 2)
-        imgh = correlate(self.img, kh, mode="nearest")
-        imgh = imgh[:, 1 : imgh.shape[1] - 1, :]
-        imgh = imgh * imgh
+    return nlevel, th, num, mask
 
-        kv = np.expand_dims(np.vstack(np.array([-0.5, 0, 0.5])), 2)
-        imgv = correlate(self.img, kv, mode="nearest")
-        imgv = imgv[1 : imgv.shape[0] - 1, :, :]
-        imgv = imgv * imgv
+def noiselevel(img, patchsize, decim, conf, itr):
+    """
+    Calculates the noise level of the input array.
+    """
+    if len(img.shape) < 3:
+        img = np.expand_dims(img, 2)
 
-        Dh = np.matrix(self.convmtx2(np.squeeze(kh, 2), self.patchsize, self.patchsize))
-        Dv = np.matrix(self.convmtx2(np.squeeze(kv, 2), self.patchsize, self.patchsize))
+    nlevel = np.ndarray(img.shape[2])
+    th = np.ndarray(img.shape[2])
+    num = np.ndarray(img.shape[2])
 
-        DD = Dh.getH() * Dh + Dv.getH() * Dv
+    kh = np.expand_dims(np.expand_dims(np.array([-0.5, 0, 0.5]), 0), 2)
+    imgh = correlate(img, kh, mode="nearest")
+    imgh = imgh[:, 1 : imgh.shape[1] - 1, :]
+    imgh = imgh * imgh
 
-        r = np.double(np.linalg.matrix_rank(DD))
-        Dtr = np.trace(DD)
+    kv = np.expand_dims(np.vstack(np.array([-0.5, 0, 0.5])), 2)
+    imgv = correlate(img, kv, mode="nearest")
+    imgv = imgv[1 : imgv.shape[0] - 1, :, :]
+    imgv = imgv * imgv
 
-        tau0 = gamma.ppf(self.conf, r / 2, scale=(2 * Dtr / r))
+    Dh = np.matrix(convmtx2(np.squeeze(kh, 2), patchsize, patchsize))
+    Dv = np.matrix(convmtx2(np.squeeze(kv, 2), patchsize, patchsize))
 
-        for cha in range(self.img.shape[2]):
-            X = view_as_windows(self.img[:, :, cha], (self.patchsize, self.patchsize))
-            X = X.reshape(
-                np.int(X.size / self.patchsize ** 2), self.patchsize ** 2, order="F"
-            ).transpose()
+    DD = Dh.getH() * Dh + Dv.getH() * Dv
 
-            Xh = view_as_windows(imgh[:, :, cha], (self.patchsize, self.patchsize - 2))
-            Xh = Xh.reshape(
-                np.int(Xh.size / ((self.patchsize - 2) * self.patchsize)),
-                ((self.patchsize - 2) * self.patchsize),
-                order="F",
-            ).transpose()
+    r = np.double(np.linalg.matrix_rank(DD))
+    Dtr = np.trace(DD)
 
-            Xv = view_as_windows(imgv[:, :, cha], (self.patchsize - 2, self.patchsize))
-            Xv = Xv.reshape(
-                np.int(Xv.size / ((self.patchsize - 2) * self.patchsize)),
-                ((self.patchsize - 2) * self.patchsize),
-                order="F",
-            ).transpose()
+    tau0 = gamma.ppf(conf, r / 2, scale=(2 * Dtr / r))
 
-            Xtr = np.expand_dims(np.sum(np.concatenate((Xh, Xv), axis=0), axis=0), 0)
+    for cha in range(img.shape[2]):
+        X = view_as_windows(img[:, :, cha], (patchsize, patchsize))
+        X = X.reshape(
+            np.int(X.size / patchsize ** 2), patchsize ** 2, order="F"
+        ).transpose()
 
-            if self.decim > 0:
-                XtrX = np.transpose(np.concatenate((Xtr, X), axis=0))
-                XtrX = np.transpose(XtrX[XtrX[:, 0].argsort(),])
-                p = np.floor(XtrX.shape[1] / (self.decim + 1))
-                p = np.expand_dims(np.arange(0, p) * (self.decim + 1), 0)
-                Xtr = XtrX[0, p.astype("int")]
-                X = np.squeeze(XtrX[1 : XtrX.shape[1], p.astype("int")])
+        Xh = view_as_windows(imgh[:, :, cha], (patchsize, patchsize - 2))
+        Xh = Xh.reshape(
+            np.int(Xh.size / ((patchsize - 2) * patchsize)),
+            ((patchsize - 2) * patchsize),
+            order="F",
+        ).transpose()
+
+        Xv = view_as_windows(imgv[:, :, cha], (patchsize - 2, patchsize))
+        Xv = Xv.reshape(
+            np.int(Xv.size / ((patchsize - 2) * patchsize)),
+            ((patchsize - 2) * patchsize),
+            order="F",
+        ).transpose()
+
+        Xtr = np.expand_dims(np.sum(np.concatenate((Xh, Xv), axis=0), axis=0), 0)
+        
+        if decim > 0:
+            XtrX = np.transpose(np.concatenate((Xtr, X), axis=0))
+            XtrX = np.transpose(XtrX[XtrX[:, 0].argsort(),])
+            p = np.floor(XtrX.shape[1] / (decim + 1))
+            p = np.expand_dims(np.arange(0, p) * (decim + 1), 0)
+            Xtr = XtrX[0, p.astype("int")]
+            X = np.squeeze(XtrX[1 : XtrX.shape[1], p.astype("int")])
+
+        # noise level estimation
+        tau = np.inf
+
+        if X.shape[1] < X.shape[0]:
+            sig2 = 0
+        else:
+            cov = (np.asmatrix(X) @ np.asmatrix(X).getH()) / (X.shape[1] - 1)
+            d = np.flip(np.linalg.eig(cov)[0], axis=0)
+            sig2 = d[0]
+
+        for i in range(1, itr):
+            # weak texture selection
+            tau = sig2 * tau0
+            p = Xtr < tau
+            Xtr = Xtr[p]
+            X = X[:, np.squeeze(p)]
 
             # noise level estimation
-            tau = np.inf
-
             if X.shape[1] < X.shape[0]:
-                sig2 = 0
-            else:
-                cov = (np.asmatrix(X) @ np.asmatrix(X).getH()) / (X.shape[1] - 1)
-                d = np.flip(np.linalg.eig(cov)[0], axis=0)
-                sig2 = d[0]
+                break
 
-            for i in range(1, self.itr):
-                # weak texture selection
-                tau = sig2 * tau0
-                p = Xtr < tau
-                Xtr = Xtr[p]
-                X = X[:, np.squeeze(p)]
+            cov = (np.asmatrix(X) @ np.asmatrix(X).getH()) / (X.shape[1] - 1)
+            d = np.flip(np.linalg.eig(cov)[0], axis=0)
+            sig2 = d[0]
 
-                # noise level estimation
-                if X.shape[1] < X.shape[0]:
-                    break
+        nlevel[cha] = np.sqrt(sig2)
+        th[cha] = tau
+        num[cha] = X.shape[1]
 
-                cov = (np.asmatrix(X) @ np.asmatrix(X).getH()) / (X.shape[1] - 1)
-                d = np.flip(np.linalg.eig(cov)[0], axis=0)
-                sig2 = d[0]
+    # clean up
+    img = np.squeeze(img)
 
-            nlevel[cha] = np.sqrt(sig2)
-            th[cha] = tau
-            num[cha] = X.shape[1]
+    return nlevel, th, num
 
-        # clean up
-        self.img = np.squeeze(self.img)
+def convmtx2(H, m, n):
+    """
+    Specialized 2D convolution matrix generation.
 
-        return nlevel, th, num
+    Parameters
+    ----------
+    H : `numpy.ndarray`
+        Input matrix.
+    m : `numpy.ndarray`
+        Rows in convolution matrix.
+    n : `numpy.ndarray`
+        Columns in convolution matrix.
 
-    def convmtx2(self, H, m, n):
-        """
-        Specialized 2D convolution matrix generation.
+    Returns
+    -------
+    T : `numpy.ndarray`
+        The new convoluted matrix.
+    """
+    s = np.shape(H)
+    m = int(m)
+    n = int(n)
+    T = np.zeros([(m - s[0] + 1) * (n - s[1] + 1), m * n])
 
-        Parameters
-        ----------
-        H : `numpy.ndarray`
-            Input matrix.
-        m : `numpy.ndarray`
-            Rows in convolution matrix.
-        n : `numpy.ndarray`
-            Columns in convolution matrix.
+    k = 0
+    for i in range((m - s[0] + 1)):
+        for j in range((n - s[1] + 1)):
+            for p in range(s[0]):
+                T[k, (i + p) * n + j : (i + p) * n + j + 1 + s[1] - 1] = H[p, :]
+    
+            k += 1
+    return T
 
-        Returns
-        -------
-        T : `numpy.ndarray`
-            The new convoluted matrix.
-        """
-        s = np.shape(H)
-        T = np.zeros([(m - s[0] + 1) * (n - s[1] + 1), m * n])
 
-        k = 0
-        for i in range((m - s[0] + 1)):
-            for j in range((n - s[1] + 1)):
-                for p in range(s[0]):
-                    T[k, (i + p) * n + j : (i + p) * n + j + 1 + s[1] - 1] = H[p, :]
-                k = k + 1
-        return T
+def weaktexturemask(img, patchsize, th):
+    """
+    Calculates the weak texture mask.
+    """
+    if len(img.shape) < 3:
+        img = np.expand_dims(img, 2)
 
-    def weaktexturemask(self):
-        """
-        Calculates the weak texture mask.
-        """
-        if len(self.img.shape) < 3:
-            self.img = np.expand_dims(self.img, 2)
+    kh = np.expand_dims(np.transpose(np.vstack(np.array([-0.5, 0, 0.5]))), 2)
+    imgh = correlate(img, kh, mode="nearest")
+    imgh = imgh[:, 1 : imgh.shape[1] - 1, :]
+    imgh = imgh * imgh
 
-        kh = np.expand_dims(np.transpose(np.vstack(np.array([-0.5, 0, 0.5]))), 2)
-        imgh = correlate(self.img, kh, mode="nearest")
-        imgh = imgh[:, 1 : imgh.shape[1] - 1, :]
-        imgh = imgh * imgh
+    kv = np.expand_dims(np.vstack(np.array([-0.5, 0, 0.5])), 1)
+    imgv = correlate(img, kv, mode="nearest")
+    imgv = imgv[1 : imgv.shape[0] - 1, :, :]
+    imgv = imgv * imgv
 
-        kv = np.expand_dims(np.vstack(np.array([-0.5, 0, 0.5])), 1)
-        imgv = correlate(self.img, kv, mode="nearest")
-        imgv = imgv[1 : imgv.shape[0] - 1, :, :]
-        imgv = imgv * imgv
+    s = img.shape
+    msk = np.zeros_like(img)
 
-        s = self.img.shape
-        msk = np.zeros_like(self.img)
-
-        for cha in range(s[2]):
-            m = view_as_windows(self.img[:, :, cha], (self.patchsize, self.patchsize))
-            m = np.zeros_like(
-                m.reshape(
-                    np.int(m.size / self.patchsize ** 2), self.patchsize ** 2, order="F"
-                ).transpose()
-            )
-
-            Xh = view_as_windows(imgh[:, :, cha], (self.patchsize, self.patchsize - 2))
-            Xh = Xh.reshape(
-                np.int(Xh.size / ((self.patchsize - 2) * self.patchsize)),
-                ((self.patchsize - 2) * self.patchsize),
-                order="F",
+    for cha in range(s[2]):
+        m = view_as_windows(img[:, :, cha], (patchsize, patchsize))
+        m = np.zeros_like(
+            m.reshape(
+                np.int(m.size / patchsize ** 2), patchsize ** 2, order="F"
             ).transpose()
+        )
 
-            Xv = view_as_windows(imgv[:, :, cha], (self.patchsize - 2, self.patchsize))
-            Xv = Xv.reshape(
-                np.int(Xv.size / ((self.patchsize - 2) * self.patchsize)),
-                ((self.patchsize - 2) * self.patchsize),
-                order="F",
-            ).transpose()
+        Xh = view_as_windows(imgh[:, :, cha], (patchsize, patchsize - 2))
+        Xh = Xh.reshape(
+            np.int(Xh.size / ((patchsize - 2) * patchsize)),
+            ((patchsize - 2) * patchsize),
+            order="F",
+        ).transpose()
 
-            Xtr = np.expand_dims(np.sum(np.concatenate((Xh, Xv), axis=0), axis=0), 0)
+        Xv = view_as_windows(imgv[:, :, cha], (patchsize - 2, patchsize))
+        Xv = Xv.reshape(
+            np.int(Xv.size / ((patchsize - 2) * patchsize)),
+            ((patchsize - 2) * patchsize),
+            order="F",
+        ).transpose()
 
-            p = Xtr < self.th[cha]
-            ind = 0
+        Xtr = np.expand_dims(np.sum(np.concatenate((Xh, Xv), axis=0), axis=0), 0)
 
-            for col in range(0, s[1] - self.patchsize + 1):
-                for row in range(0, s[0] - self.patchsize + 1):
-                    if p[:, ind]:
-                        msk[
-                            row : row + self.patchsize - 1,
-                            col : col + self.patchsize - 1,
-                            cha,
-                        ] = 1
-                    ind = ind + 1
+        p = Xtr < th[cha]
+        ind = 0
 
-        # clean up
-        self.img = np.squeeze(self.img)
+        for col in range(0, s[1] - patchsize + 1):
+            for row in range(0, s[0] - patchsize + 1):
+                if p[:, ind]:
+                    msk[
+                        row : row + patchsize - 1,
+                        col : col + patchsize - 1,
+                        cha,
+                    ] = 1
+                ind = ind + 1
 
-        return np.squeeze(msk)
+    # clean up
+    img = np.squeeze(img)
+
+    return np.squeeze(msk)
